Migration CRA to Vite
