--- .vscode/settings.json
@@ -1,14 +1,0 @@
-{
-	"editor.defaultFormatter": "esbenp.prettier-vscode",
-	"editor.formatOnSave": true,
-	"[javascript]": {
-		"editor.formatOnSave": false
-	},
-	"[javascriptreact]": {
-		"editor.formatOnSave": false
-	},
-	"editor.codeActionsOnSave": {
-		"source.fixAll": true
-	},
-	"eslint.alwaysShowStatus": true
-}
