+++ package.json
@@ -2,53 +2,34 @@
 	"name": "todo-app",
 	"version": "0.1.0",
 	"private": true,
+	"type": "module",
+	"scripts": {
+		"dev": "vite",
+		"build": "vite build",
+		"lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+		"preview": "vite preview"
+	},
 	"dependencies": {
 		"@fontsource/poppins": "^4.5.0",
 		"@reduxjs/toolkit": "^1.7.1",
 		"date-fns": "^2.28.0",
 		"framer-motion": "^5.5.1",
-		"node-sass": "^7.0.0",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
 		"react-hot-toast": "^2.2.0",
 		"react-icons": "^4.3.1",
 		"react-redux": "^7.2.6",
-		"react-scripts": "5.0.0",
-		"uuid": "^8.3.2"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject"
-	},
-	"eslintConfig": {
-		"extends": [
-			"wesbos"
-		],
-		"rules": {
-			"import/no-extraneous-dependencies": [
-				"error",
-				{
-					"devDependencies": true
-				}
-			],
-			"react/prop-types": 0
-		}
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
+		"uuid": "^8.3.2",
 	},
 	"devDependencies": {
-		"eslint-config-wesbos": "^3.0.2"
+		"sass": "^1.69.5",
+		"@vitejs/plugin-react": "^4.2.0",
+		"browserslist-to-esbuild": "^1.2.0",
+		"eslint": "^8.53.0",
+		"eslint-plugin-react": "^7.33.2",
+		"eslint-plugin-react-hooks": "^4.6.0",
+		"eslint-plugin-react-refresh": "^0.4.4",
+		"vite": "^5.0.0",
+		"vite-jsconfig-paths": "^2.0.1"
 	}
 }
