+++ src/app/store.js
@@ -1,8 +1,8 @@
-import { configureStore } from '@reduxjs/toolkit';
-import todoReducer from '../slices/todoSlice';
+import { configureStore } from "@reduxjs/toolkit";
+import todoReducer from "slices/todoSlice";
 
 export const store = configureStore({
-  reducer: {
-    todo: todoReducer,
-  },
+	reducer: {
+		todo: todoReducer,
+	},
 });
