+++ src/styles/GlobalStyles.css
@@ -5,6 +5,7 @@
 	margin: 0;
 	padding: 0;
 }
+
 :root {
 	--primaryPurple: #646ff0;
 	--black-1: #646681;
@@ -17,18 +18,22 @@
 	--black: black;
 	--white: white;
 }
+
 html {
 	font-size: 10px;
 }
+
 body {
 	font-family: 'Poppins', sans-serif;
 	width: 100%;
 	min-height: 100vh;
 	background-color: var(--bg-1);
 }
+
 * {
 	font-family: 'Poppins', sans-serif;
 }
+
 .container {
 	width: 90%;
 	max-width: 1200px;
