+++ src/styles/modules/app.module.scss
@@ -5,12 +5,14 @@
 	width: 100%;
 	margin: 0 auto;
 }
+
 .appHeader {
 	display: flex;
 	align-items: center;
 	justify-content: space-between;
 	height: 60px;
 }
+
 .content__wrapper {
 	background-color: var(--bg-2);
 	padding: 2rem;
@@ -19,6 +21,7 @@
 		padding: 1.5rem;
 	}
 }
+
 .emptyText {
 	// display: inline-block;
 	font-size: 1.6rem;
