+++ src/styles/modules/title.module.scss
@@ -1,4 +1,5 @@
 @import '../mixins';
+
 .title {
 	display: inline-block;
 	width: 100%;
