+++ src/styles/modules/todoItem.module.scss
@@ -8,16 +8,19 @@
 	background: var(--white);
 	margin-bottom: 1.5rem;
 	border-radius: 4px;
+	
 	&:last-child {
 		margin-bottom: 0;
 	}
 }
+
 .todoDetails {
 	display: flex;
 	align-items: center;
 	justify-content: flex-start;
 	gap: 1rem;
 }
+
 .svgBox {
 	flex-basis: 25px;
 	flex-shrink: 0;
@@ -29,9 +32,11 @@
 	padding: 5px;
 	cursor: pointer;
 	transition: 0.3s ease background-color;
+	
 	&:hover {
 		background-color: var(--grey-2);
 	}
+	
 	svg {
 		width: 100%;
 		height: 100%;
@@ -41,22 +46,25 @@
 		justify-content: center;
 	}
 }
+
 .texts {
 	display: flex;
 	flex-direction: column;
 	overflow: hidden;
 }
+
 .todoText {
 	word-break: break-all;
 	font-weight: 500;
 	font-size: 1.4rem;
 	color: var(--black-2);
-
+	
 	&--completed {
 		text-decoration: line-through;
 		opacity: 0.7;
 	}
 }
+
 .time {
 	display: block;
 	font-size: 1.2rem;
@@ -64,12 +72,14 @@
 	margin-top: -0.2rem;
 	color: var(--black-2);
 }
+
 .todoActions {
 	display: flex;
 	align-items: center;
 	justify-content: center;
 	gap: 1rem;
 }
+
 .icon {
 	font-size: 2rem;
 	padding: 0.5rem;
@@ -81,6 +91,7 @@
 	justify-content: center;
 	cursor: pointer;
 	transition: 0.3s ease background-color;
+	
 	&:hover {
 		background-color: var(--gray-2);
 	}
