+++ src/utils/getClasses.js
@@ -1,5 +1,5 @@
 export const getClasses = (classes) =>
-  classes
-    .filter((item) => item !== '')
-    .join(' ')
-    .trim();
+	classes
+	.filter((item) => item !== "")
+	.join(" ")
+	.trim();
